refactor : 기능별 명확성을 위해 서비스, 컨트롤러 명칭 변경
